+++ ChangeLog.md
@@ -37,9 +37,8 @@ Format is `version (*dd-mm-yyyy*)`
  * Added `.github/SUPPORT.md` for SUPPORT
  * Added `debug_info` in tests
  * Added removing null character in user input
- * Added whether to `escape` and to `preventXSS` in configuration
+ * Added whether to `preventXSS` in configuration
  * Added `sanitizeArray()` function - sanitizes an array
- * Added `escape()` function - escape a string
  * Added `sanitize("message", $message);`, for strings that contains EOL (END OF LINE) [commonly used for contact forms]
  * Added Sanitize HTML code using `$sanitizer->HTML($code)`
  * Added loading config from Ini file using `configFromIni("/path/to/src/config.ini");`